Clean-up of logging. Now a test can override description(...) to supply a custom description for the logging.
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@3993 e85b7adc-8362-4630-8c63-7469d557c915

+++ test/test.cc
@@ -76,7 +76,8 @@ main(int argc, char** argv) {
   unsigned int testCount = 0;
   for ( Test* t = Test::tests(); t != NULL; t = t->next() )
     testCount++;
-  int digits = std::ceil(std::log10(static_cast<double>(testCount)));
+  int digits = 
+    static_cast<int>(std::ceil(std::log10(static_cast<double>(testCount))));
   
   unsigned int counter = 1;
   Test* t = Test::tests();
